requirements.txt
fastapi~=0.112.0
pydantic~=2.8.2
requests~=2.32.3
python-dotenv~=1.0.1